Add core repo files [OPTIONAL] Welcome to the first commit!
This step is meant to get your git repo up and running in the correct
way and in the correct location (/wp-content). The repo root is in
wp-content so we can separate our plugin code (always runs regardless of
the theme) and the theme (how the content is output). If you don't
intend to use git, this step can be skipped.

See the README.md file for this repo for information on how these files
were created.