Add core theme files This step is a critical one as it makes the theme able to be activated in wp-admin.
An important thing to note here that could be easily overlooked is the
directory for your custom theme. The directory name should be succinct
and only use lower-case letters, numbers, and dashes.

Make sure to read the first comment block in style.css to understand
what goes there.

Once this step is complete in your own install, go to wp-admin >
Appearance > Themes and make sure it appears there. If it does not, you
should see a message at the bottom of that screen with more information.

Click Activate and you're ready for the next step!