Add important but not required theme files
WP themes don't require a functions.php or screenshot.png file to work
but if you're adding any custom functionality, which we will, and if you
don't want your Manage Themes page to look broken, which we don't, then
these are a good thing to have early on.

As you'll see in a later commit, functions.php has a few very specific
duties and is not the dumping ground for PHP snippets. The same goes for
themes, not all code belongs in the theme. Keeping walking through the
commits in this repo and you should not be left feeling lost where
things should go.

The Codex has guidelines for appropriate size and formats for
screenshots
but, to save you a click, go with a 1200 x 900px PNG. More here:

https://codex.wordpress.org/Theme_Development#Screenshot